asdf

+++ backend/src/main/java/com/ssafy/gaese/domain/algorithm/application/AlgoService.java
@@ -394,7 +394,7 @@ public class AlgoService {
             algoCount++;
         }
 
-        int charId=21;
+        long charId=21;
         if(oneCount>0 && !userCharacterRepository.findByUserAndCharacters(user,characters.get(charId)).isPresent())
         {
             userCharacterSet(user,charId,characters);
